rock_ci_setupdev: delete dev/autoproj and always bootstrap
The configuration files do not contain the version control information
(if they did, we could not share them), so the best way to update the
conf is to delete dev/autoproj (and only dev/autoproj) and bootstrap
again. Since the rest (including the gems) is already there, it is
pretty cheap to do so.